Таск закончен.

+++ app/services/db/models.py
@@ -1,6 +1,5 @@
+from sqlalchemy import Column, Float, Integer, String, DECIMAL
 from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy import Column, Integer, Float, String
-
 
 Base = declarative_base()
 metadata = Base.metadata
@@ -10,8 +9,8 @@ class User(Base):
     __tablename__ = "users"
     id = Column(Integer, primary_key=True)
     name = Column(String(128), unique=True)
-    x = Column(Float, default=False)
-    y = Column(Float, default=False)
+    x = Column(DECIMAL, default=False)
+    y = Column(DECIMAL, default=False)
 
 
 UserTable = User.__table__
